// scripts/types/database/public/EspeceProtegeeModification.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.espece_protegee_modification */
export type EspeceProtegeeModificationCdRef = string & { __brand: 'public.espece_protegee_modification' };

/** Represents the table public.espece_protegee_modification */
export default interface EspeceProtegeeModification {
  cd_ref: EspeceProtegeeModificationCdRef;

  classification: string | null;

  noms_scientifiques: string[] | null;

  noms_vernaculaires: string[] | null;

  cd_type_statuts: string[] | null;

  espece_ministerielle: boolean | null;

  espece_cnpn: boolean | null;

  exclu: boolean;

  modifie_par: string | null;

  created_at: Date;

  updated_at: Date;
}

/** Represents the initializer for the table public.espece_protegee_modification */
export interface EspeceProtegeeModificationInitializer {
  cd_ref: EspeceProtegeeModificationCdRef;

  classification?: string | null;

  noms_scientifiques?: string[] | null;

  noms_vernaculaires?: string[] | null;

  cd_type_statuts?: string[] | null;

  espece_ministerielle?: boolean | null;

  espece_cnpn?: boolean | null;

  /** Default value: false */
  exclu?: boolean;

  modifie_par?: string | null;

  /** Default value: CURRENT_TIMESTAMP */
  created_at?: Date;

  /** Default value: CURRENT_TIMESTAMP */
  updated_at?: Date;
}

/** Represents the mutator for the table public.espece_protegee_modification */
export interface EspeceProtegeeModificationMutator {
  cd_ref?: EspeceProtegeeModificationCdRef;

  classification?: string | null;

  noms_scientifiques?: string[] | null;

  noms_vernaculaires?: string[] | null;

  cd_type_statuts?: string[] | null;

  espece_ministerielle?: boolean | null;

  espece_cnpn?: boolean | null;

  exclu?: boolean;

  modifie_par?: string | null;

  created_at?: Date;

  updated_at?: Date;
}